+++ example/page_manager.dart
@@ -7,22 +7,28 @@ class PageManager {
     //you are able to check whether all required arguments are available
     var hasIsErrorOccurred = args.hasArgument("isErrorOccurred");
     var hasErrorMsg = args.hasArgument("errorMsg");
-    var hasPrintError = args.hasArgument("printError");
 
-    if(hasIsErrorOccurred && hasErrorMsg && hasPrintError) {
+    if(hasIsErrorOccurred && hasErrorMsg) {
 
       //get and cast arguments
       var isErrorOccurred = args["isErrorOccurred"] as bool;
       var errorMsg = args["errorMsg"] as String;
-      var printError = args["printError"] as Function;
 
       //do whatever you want with it ;)
       if(isErrorOccurred) {
-        printError(errorMsg);
-      } else {
-        args.printSuccess(args.successMsg);
+        print(errorMsg);
+      }
+
+      //you can also access arguments by using the dot notation
+      else {
+        print(args.successMsg);
       }
     }
+  });
+
+  //create another slot which is invoked when a page is loaded
+  Slot _onPageLoadedTwo = new Slot((args) {
+    print("hey!");
   });
 
   PageManager() {
@@ -30,8 +36,9 @@ class PageManager {
     //the "Page_OnLoaded" signal is emitted
     //the _onPageLoaded Slot is invoked.
     Connect
-    .signal(Page.ON_LOADED)
-    .to(_onPageLoaded);
+      .signal(Page.ON_LOADED)
+      .to(_onPageLoadedTwo)
+      .to(_onPageLoaded);
   }
 
   //loads all existing pages
@@ -43,7 +50,8 @@ class PageManager {
   //disconnects signal from slot
   void destroy() {
     Disconnect
-    .signal(Page.ON_LOADED)
-    .from(_onPageLoaded);
+      .signal(Page.ON_LOADED)
+      .from(_onPageLoadedTwo)
+      .from(_onPageLoaded);
   }
 }